Fixed LFO delay LUT and interpetation of "Vibrato Delay" parameter
The existing LFO delay lookup table was based on the premise that
changes to "Vibrato Delay" parameter had the same effect as changing
the LFO1/2 Delay ROM values, but this turned out to be wrong.

With this change we now have a good approximation of LFO delay times,
including changes via the "Vibrato Delay" parameter.