Add support for CPack generation of TGZ, DEB, RPM and Windows installer
On linux we now support auto generation of .deb and .rpm package files.
A desktop file is also added to make the emusc application show up in Linux
desktop environments.

On Windows the NSIS installer creates an installation program including
all files needed to run emusc, and develope on libemusc, on any modern
version of windows.

A pkg-config file was added to libemusc to simplify external use of the
library.