Improve TVA envelope time calculations and exponential decay
A deep dive into volume (TVA) envelope time calculations revealed new insight.
First, the 5 phase envelopes used in the sound canvas are actually Attack1,
Attack2, Decay1, Decay2 and Release. This has implications for which phases
that are affected by different parameters. Names are updated in the source
code.

Partial definition byte 81 is identified to be "TVA Envelope Time Key Adjust"
for Attack and Decay phases, while byte 82 is the same for the release phase.
This has been implemented, but with a somewhat crude simplifications in the
case of value=1.

Partial definition byte 83 is identidifed to be "TVA Envelope Time Key Follow"
for Attack and Decay phases, while byte 84 is the same for the release phase.
Added a pretty accurate approximation of this function.

And finally fixed some old bugs and improved debug output in the Envelope
class.