Fix regression in exponential decay envelope time calculation
Previous commit containing TVA timing fixes was missing a commit log about
changes to the logarithmic envelope time decay. Turned out that the Sound
Canvas is not actually using logarithmic decay, but rather exponential decay.

A relatively complex approximation of the exponential function was added, but it
had a bug that made the new function always assume that the end value would be
zero. This commit fixes this issue so that any start and end levels can be used.