Improve support for TVF filters
It was just discovered that there is a control byte in the partial definition
for specifying low pass filter, high pass filter or disable TVF. This commit
add support for setting the correct filter type or disabling it.

The correct calculation for the cutoff frequency in the base filter has also
been identified and added in this commit.

Based on the results from these improvements it seems empirically proven that
the Sound Canvas is using 2nd. order biquad filters for TVF.