Fix TVA Envelope Key Follow T1-T4 and T5 when Preset=1
It turned out that calculating the correct TVA Envelope Key Follow was a few
notches more complex than first anticipated when Preset=1. In total 4 different
lookup tables are used to get the correct envelope length corrections.

Pitch and TVF Envelope Key Follow does not have the Preset option and are not
affected by this patch.