Code refactoring to improve usage of lookup tables in the PROGROM
The PROGROM contains a significant number of LUTs (lookup tables). It turns
out that one specific LUT in the ROM is actually a memory map for the all the
other LUTs (well, except for the velocity curves). This patch removes all
manual work to access these LUTs individually and rather use the mapping LUT
to dynamically find other LUTs.

This patch also reimplements the calculations done using these PROGROM LUTs
similarly to what is done in hardware (based on disassembling the binary code
in the CPUROM). This gives exact same audio calculations as on the orignal
hardware, and in the most CPU efficient way.

Another benefit from this code restructuring is that EmuSC now automatically
changes behaviour for other models, such as the SC55mkII, which has quite a
few differences in the PROGROM LUTs.