Rewritten waveform (LFO) generator class and LFO dialog
Based on inspection of the original machine code in the CPU ROM the waveform
generation code has been rewritten to match the same logic. This makes the
source code more low level, something that might be harder to read and less
generic, but that is very fast and produces identical waveforms, phase shifts,
delays and fade effects as the original hardware synth.

The LFO dialog have been rewritten to give more precise details about the LFOs.

Also added generic Part changed callback for frontends to automatically update
when changes are done through e.g. SysEx messages to a Part.