Code rewrite of dynamic TVA parameters
Based on CPU instructions on the CPU ROM the TVA calculations for dynamic
parameters, such as expression, volume controls, controller inputs and LFO
modulations, have been rewritten to use exact same corrections and integer
arithmetic as the original hardware. This makes the code more low level and
more difficult to read, but also gives -identical- output as the hardware
synth - at a minimum of CPU usage.

This commit also fixes a crash when pan was set to -64, and it re-enables LFO
modulation (tremolo) for TVA.